Rename sequencer file for consistency with fullnode / prover
